+++ lib/src/ast/css/modifiable/style_rule.dart
@@ -21,12 +21,13 @@ final class ModifiableCssStyleRule extends ModifiableCssParentNode
 
   final SelectorList originalSelector;
   final FileSpan span;
+  final bool fromPlainCss;
 
   /// Creates a new [ModifiableCssStyleRule].
   ///
   /// If [originalSelector] isn't passed, it defaults to [_selector.value].
   ModifiableCssStyleRule(this._selector, this.span,
-      {SelectorList? originalSelector})
+      {SelectorList? originalSelector, this.fromPlainCss = false})
       : originalSelector = originalSelector ?? _selector.value;
 
   T accept<T>(ModifiableCssVisitor<T> visitor) =>