+++ lib/src/parse/stylesheet.dart
@@ -56,6 +56,11 @@ abstract class StylesheetParser extends Parser {
   /// Whether the parser is currently within a parenthesized expression.
   var _inParentheses = false;
 
+  /// Whether the parser is currently within an expression.
+  @protected
+  bool get inExpression => _inExpression;
+  var _inExpression = false;
+
   /// A map from all variable names that are assigned with `!global` in the
   /// current stylesheet to the nodes where they're defined.
   ///
@@ -319,10 +324,6 @@ abstract class StylesheetParser extends Parser {
   ///   parsed as a selector and never as a property with nested properties
   ///   beneath it.
   Statement _declarationOrStyleRule() {
-    if (plainCss && _inStyleRule && !_inUnknownAtRule) {
-      return _propertyOrVariableDeclaration();
-    }
-
     // The indented syntax allows a single backslash to distinguish a style rule
     // from old-style property syntax. We don't support old property syntax, but
     // we do support the backslash because it's easy to do.
@@ -395,10 +396,7 @@ abstract class StylesheetParser extends Parser {
     }
 
     var postColonWhitespace = rawText(whitespace);
-    if (lookingAtChildren()) {
-      return _withChildren(_declarationChild, start,
-          (children, span) => Declaration.nested(name, children, span));
-    }
+    if (_tryDeclarationChildren(name, start) case var nested?) return nested;
 
     midBuffer.write(postColonWhitespace);
     var couldBeSelector =
@@ -434,12 +432,8 @@ abstract class StylesheetParser extends Parser {
       return nameBuffer;
     }
 
-    if (lookingAtChildren()) {
-      return _withChildren(
-          _declarationChild,
-          start,
-          (children, span) =>
-              Declaration.nested(name, children, span, value: value));
+    if (_tryDeclarationChildren(name, start, value: value) case var nested?) {
+      return nested;
     } else {
       expectStatementSeparator();
       return Declaration(name, value, scanner.spanFrom(start));
@@ -544,29 +538,34 @@ abstract class StylesheetParser extends Parser {
     }
 
     whitespace();
-
-    if (lookingAtChildren()) {
-      if (plainCss) {
-        scanner.error("Nested declarations aren't allowed in plain CSS.");
-      }
-      return _withChildren(_declarationChild, start,
-          (children, span) => Declaration.nested(name, children, span));
-    }
+    if (_tryDeclarationChildren(name, start) case var nested?) return nested;
 
     var value = _expression();
-    if (lookingAtChildren()) {
-      if (plainCss) {
-        scanner.error("Nested declarations aren't allowed in plain CSS.");
-      }
-      return _withChildren(
-          _declarationChild,
-          start,
-          (children, span) =>
-              Declaration.nested(name, children, span, value: value));
+    if (_tryDeclarationChildren(name, start, value: value) case var nested?) {
+      return nested;
     } else {
       expectStatementSeparator();
       return Declaration(name, value, scanner.spanFrom(start));
     }
+  }
+
+  /// Tries parsing nested children of a declaration whose [name] has already
+  /// been parsed, and returns `null` if it doesn't have any.
+  ///
+  /// If [value] is passed, it's used as the value of the peroperty without
+  /// nesting.
+  Declaration? _tryDeclarationChildren(
+      Interpolation name, LineScannerState start,
+      {Expression? value}) {
+    if (!lookingAtChildren()) return null;
+    if (plainCss) {
+      scanner.error("Nested declarations aren't allowed in plain CSS.");
+    }
+    return _withChildren(
+        _declarationChild,
+        start,
+        (children, span) =>
+            Declaration.nested(name, children, span, value: value));
   }
 
   /// Consumes a statement that's allowed within a declaration.
@@ -1686,7 +1685,9 @@ abstract class StylesheetParser extends Parser {
     }
 
     var start = scanner.state;
+    var wasInExpression = _inExpression;
     var wasInParentheses = _inParentheses;
+    _inExpression = true;
 
     // We use the convention below of referring to nullable variables that are
     // shared across anonymous functions in this method with a trailing
@@ -2039,11 +2040,13 @@ abstract class StylesheetParser extends Parser {
       _inParentheses = wasInParentheses;
       var singleExpression = singleExpression_;
       if (singleExpression != null) commaExpressions.add(singleExpression);
+      _inExpression = wasInExpression;
       return ListExpression(commaExpressions, ListSeparator.comma,
           scanner.spanFrom(beforeBracket ?? start),
           brackets: bracketList);
     } else if (bracketList && spaceExpressions != null) {
       resolveOperations();
+      _inExpression = wasInExpression;
       return ListExpression(spaceExpressions..add(singleExpression_!),
           ListSeparator.space, scanner.spanFrom(beforeBracket!),
           brackets: true);
@@ -2054,6 +2057,7 @@ abstract class StylesheetParser extends Parser {
             ListSeparator.undecided, scanner.spanFrom(beforeBracket!),
             brackets: true);
       }
+      _inExpression = wasInExpression;
       return singleExpression_!;
     }
   }