+++ pkg/sass_api/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 10.0.0
+
+* Remove the `allowPlaceholders` argument from `SelectorList.parse()`. Instead,
+  it now has a more generic `plainCss` argument which tells it to parse the
+  selector in plain CSS mode.
+
+* Rename `SelectorList.resolveParentSelectors` to `SelectorList.nestWithin`.
+
+## 9.5.0
+
+* No user-visible changes.
+
+## 9.4.2
+
+* No user-visible changes.
+
 ## 9.4.1
 
 * No user-visible changes.